+++ CHANGELOG.md
@@ -1,6 +1,9 @@
 Blackcube Admin - CHANGELOG 
 ===========================
 
+ * Fix: Fix `radiolist` and `dropdownlist` typo `radioList` and `dropdownList`
+ * Upd: cleanup `Html` helper
+
 Release 3.0.6 October, 8th 2023 <pgaultier@redcat.io>
 ----------------------------------------------------