+++ CHANGELOG.md
@@ -1,10 +1,13 @@
 Blackcube Admin - CHANGELOG 
 ===========================
 
- * Upd: Adding export system
+Release 3.0.7 October, 16th 2023 <pgaultier@redcat.io>
+----------------------------------------------------
+
  * Upd: Update dependencies
+ * Upd: Add virtual columns for filtering in BLoc
+ * Upd: Adding nofollow to external links with Quill helper (`core`)
  * Fix: Fix `radiolist` and `dropdownlist` typo `radioList` and `dropdownList`
- * Upd: cleanup `Html` helper
 
 Release 3.0.6 October, 8th 2023 <pgaultier@redcat.io>
 ----------------------------------------------------